PPTGeneration/services/pptgeneration/requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0

# HTTP client
httpx>=0.24.0,<1.0.0

# Vector database client
qdrant-client>=1.6.0,<2.0.0

# Email parsing
extract-msg>=0.41.0,<1.0.0

# HTML processing
beautifulsoup4>=4.12.0,<5.0.0

# Date utilities
python-dateutil>=2.8.0,<3.0.0

# Data validation
pydantic>=2.0.0,<3.0.0

# PowerPoint generation
python-pptx>=0.6.21,<1.0.0

# System monitoring (optional)
psutil>=5.9.0,<6.0.0

# Development and testing (optional)
pytest>=7.0.0,<8.0.0
pytest-django>=4.5.0,<5.0.0
